Removed simplify.py and moved simplify_in_constant() to symbol.py
Also removed import of symplify in __init__.py
